# apidoc.json
{
  "swagger": "2.0",
  "info": {
    "description": "Rebbl.net API",
    "version": "2",
    "title": "Rebbl.net API"
  },
  "host": "rebbl.net",
  "basePath": "/api/v2",
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/league": {
      "get": {
        "tags": [
          "league"
        ],
        "summary": "Find all leagues",
        "description": "Retrieve a list of all configures leagues in ReBBL",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/League"
              }
            }
          },
          "400": {
            "description": "something something error"
          }
        }
      }
    },
    "/league/{leagueName}/Seasons": {
      "get": {
        "tags": [
          "league"
        ],
        "summary": "Find seasons by League Name",
        "description": "Returns a single league",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "leagueName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Season"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        }
      }
    },
    "/league/{leagueName}/{season}/slim/{round}": {
      "get": {
        "tags": [
          "league"
        ],
        "summary": "Get match results from entire league per season and round",
        "description": "Returns a single league",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "leagueName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "season",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "round",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/LeagueSlim"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        }
      }
    },
...
  },
  "definitions": {
    "League": {
      "type": "string"
    },
    "Season": {
      "type": "string"
    },
...
    "LeagueSlim": {
      "type": "object",
      "properties": {
        "division":{
          "type":"string"
        },
        "round": {
          "type": "integer"
        },
        "match_uuid": {
          "type": "string"
        },
        "homeCoachId" : {"type":"integer"},
        "homeCoachName" : {"type":"string"},
        "homeTeamId" : {"type":"integer"},
        "homeTeamName" : {"type":"string"},
        "homeTeamRace" : {"type":"string"},
        "homeScore" : {"type":"integer"},
        "awayCoachId" : {"type":"integer"},
        "awayCoachName" : {"type":"string"},
        "awayTeamId" : {"type":"integer"},
        "awayTeamName" : {"type":"string"},
        "awayTeamRace" : {"type":"string"},
        "awayScore" : {"type":"integer"}
      }
    },
...
  }
}